move constant to api
