+++ src/api/java/com/minecolonies/api/utils/constant/IToolType.java
@@ -1,4 +1,4 @@
-package com.minecolonies.coremod.util.constants;
+package com.minecolonies.api.util.constant;
 
 public interface IToolType
 {
