+++ src/api/java/com/minecolonies/api/utils/constant/ToolType.java
@@ -1,4 +1,4 @@
-package com.minecolonies.coremod.util.constants;
+package com.minecolonies.api.util.constant;
 
 import java.util.HashMap;
 import java.util.Map;
