+++ src/api/java/com/minecolonies/api/utils/constant/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains groups of constants we use at multiple places.
  */
-package com.minecolonies.coremod.util.constants;
+package com.minecolonies.api.util.constant;
