+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/guard/EntityAIMeleeGuard.java
@@ -4,7 +4,7 @@ import com.minecolonies.compatibility.Compatibility;
 import com.minecolonies.coremod.colony.jobs.JobGuard;
 import com.minecolonies.coremod.entity.ai.util.AIState;
 import com.minecolonies.coremod.entity.ai.util.AITarget;
-import com.minecolonies.coremod.util.constants.ToolType;
+import com.minecolonies.api.util.constant.ToolType;
 import com.minecolonies.coremod.util.InventoryFunctions;
 import com.minecolonies.coremod.util.ItemStackUtils;
 import net.minecraft.enchantment.EnchantmentHelper;
