+++ src/main/java/com/minecolonies/coremod/util/InventoryUtils.java
@@ -1,6 +1,6 @@
 package com.minecolonies.coremod.util;
 
-import com.minecolonies.coremod.util.constants.IToolType;
+import com.minecolonies.api.util.constant.IToolType;
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
